perform foreach iteration on a custom linked list by implementing  IEnumerator
this code created your own linked list (does not use the inbuilt c# linked list) and the user can perform for each operation on this List. The list is a generic type and can create a list of any data type. the code also shows the use of indexers and how to implement IEnumerator which is inherited from the IEnumerable